decoupled input layer

+++ src/main/java/com/avnishsrivastava/input/IUserInput.java
@@ -6,5 +6,5 @@ public interface IUserInput {
     *
     */
 
-    long getLong();
+    int getInt();
 }
